style(router): move sample application source
Moved saple application source into app folder in preparation for Cypress tests.
