+++ packages/router/test/e2e/doc-example/app/src/components/authors/author-details.ts
