+++ packages/router/test/e2e/doc-example/app/webpack.config.js
@@ -1,0 +1,27 @@
+const HtmlWebpackPlugin = require('html-webpack-plugin');
+
+module.exports = function(env, { mode }) {
+  const production = mode === 'production';
+  return {
+    mode: production ? 'production' : 'development',
+    entry: './src/startup.ts',
+    devtool: production ? 'source-map' : 'eval-source-map',
+    resolve: {
+      extensions: ['.ts', '.js'],
+      modules: ['src', 'node_modules']
+    },
+    devServer: {
+      port: 9000,
+      historyApiFallback: true,
+      open: true,
+      lazy: false
+    },
+    module: {
+      rules: [
+        { test: /\.ts$/i, loader: 'ts-loader' },
+        { test: /\.html$/i, loader: 'html-loader' }
+      ]
+    },
+    plugins: [new HtmlWebpackPlugin({ template: 'index.ejs' })]
+  }
+}
